paskaita

+++ src/main/java/example_4/Product.java
@@ -3,9 +3,9 @@ package example_4;
 public class Product {
     public String name;
     public int price;
-    public int weight;
+    public double weight;
 
-    public Product(String name, int price, int weight) {
+    public Product(String name, int price, double weight) {
         this.name = name;
         this.price = price;
         this.weight = weight;
